gke/ib-gateway/healthcheck/requirements.txt
falcon==2.0.0
google-cloud-logging==1.14.0
gunicorn==20.0.4
ib-insync==0.9.58
